+++ go/libraries/doltcore/sqle/database.go
@@ -52,16 +52,16 @@ type Database struct {
 	requestedName string
 	ddb           *doltdb.DoltDB
 	rsr           env.RepoStateReader
-	rsw           env.RepoStateWriter
-	gs            globalstate.GlobalState
-	editOpts      editor.Options
+	rsw      env.RepoStateWriter
+	gs       dsess.GlobalStateImpl
+	editOpts editor.Options
 	revision      string
 	revType       dsess.RevisionType
 }
 
 var _ dsess.SqlDatabase = Database{}
 var _ dsess.RevisionDatabase = Database{}
-var _ dsess.GlobalStateProvider = Database{}
+var _ globalstate.GlobalStateProvider = Database{}
 var _ sql.CollatedDatabase = Database{}
 var _ sql.Database = Database{}
 var _ sql.StoredProcedureDatabase = Database{}
@@ -134,7 +134,7 @@ func (db Database) DoltDatabases() []*doltdb.DoltDB {
 
 // NewDatabase returns a new dolt database to use in queries.
 func NewDatabase(ctx context.Context, name string, dbData env.DbData, editOpts editor.Options) (Database, error) {
-	globalState, err := globalstate.NewGlobalStateStoreForDb(ctx, name, dbData.Ddb)
+	globalState, err := dsess.NewGlobalStateStoreForDb(ctx, name, dbData.Ddb)
 	if err != nil {
 		return Database{}, err
 	}
@@ -211,7 +211,7 @@ func (db Database) DbData() env.DbData {
 	}
 }
 
-func (db Database) GetGlobalState() globalstate.GlobalState {
+func (db Database) GetGlobalState() dsess.GlobalStateImpl {
 	return db.gs
 }
